refactor(prerender): remove dead shared.js and rename html-comparator util
Delete src/prerender/utils/shared.js — no source file imports it; its
agentic-sheet functions migrated to src/utils/agentic-urls.js. Remove
all stale esmock stubs and unit tests referencing it from handler.test.js.

Rename utils/html-comparator.js → utils/html-analyzer.js to eliminate
the confusing same-name sibling pattern with the root html-comparator.js.
The name now matches the wrapped @adobe/spacecat-shared-html-analyzer package.
Update the one source import and all test esmock paths.

595 tests, 0 failures.
